ajusted step function
