--- perceptron/bin/Debug/netcoreapp2.0/perceptron.runtimeconfig.dev.json
@@ -1,9 +1,0 @@
-{
-  "runtimeOptions": {
-    "additionalProbingPaths": [
-      "C:\\Users\\jpsneto\\.dotnet\\store\\|arch|\\|tfm|",
-      "C:\\Users\\jpsneto\\.nuget\\packages",
-      "C:\\Program Files\\dotnet\\sdk\\NuGetFallbackFolder"
-    ]
-  }
-}
